QPPA-3759 - fixed definitions not resolving (#339)
* updated `eslint-plugin-react` to current version

* fixed lint error - removed unnecessary semicolon

* updated `measureSpecification` def to validate correctly, updated 2017 measure schema id link, added file exists check

+++ test/scripts/measures/2018/import-quality-measures-spec.js
@@ -58,7 +58,9 @@ describe(year + ' import quality measures', function() {
 
 after(function() {
   try {
-    fs.unlinkSync(outputPath);
+    if (fs.existsSync(outputPath)) {
+      fs.unlinkSync(outputPath);
+    }
   } catch (err) {
     console.log('Warning: ', outputPath, ' should have been deleted but was not.');
   }
